WIP
i think this is totally not the way that it should work.

if we leak these into the blobs system it will be totally swampt and
unusable.

i'd still wager for either a new feed suffix type or _at the very least_
use new suffix like .box and just post a string instead of a message
with a referenced blob inside of it.

the nice way of the above is, that the _content hash_ to be deleted
becomes the message hash and makes sure you can't delete other peoples
msgs with the same content (or removes the need to track these unused
references)

i suspect privgroups will introduce a new one anyway so we might just
make their life easier.